Modify security policy for the org
The security policy was originally written for the .github repo
specifically, but because .github repo is special, was applying as the
security policy for the entire org.

This rewords the security policy to be more general (so that it applies
_correctly_ to all nodenv repos) and removes all relative links (which
don't resolve correctly when the policy is rendered on other pages).