Allow releases to opt out of vMajor
This one defaults to off, though, since pushing these vMajor refs is
limited to actions and workflow repositories.

Also rename the job since we're not pushing tags right now (pushing
branches).

Also drop the ref_type check because that should really be checked by
the caller. Or if we want to check it in the called workflow, we should
be checking it on _every_ job. Meh.