Don't error github release if already exists
Common setup for running this release is on tag events.
But if the github release UI is used to create the tag, then the release
already exists. In that case, we should just exit without error.

Ideally we'd "skip" or "cancel" instead of exiting with success. But if
the end result is "release exists", then we're technically successfull.
Still better to succeed than fail. Future improvement can exit as
canceled or skipped or something if that's possible.

fixes #85