Extract Major Version job
This job only applies to workflows and github actions. The majority of
the repos using these workflows are npm packages, which don't need the
job. Instead of having the job be included and off-by-default, better to
just extract it to its own workflow that can be used in the few places
necessary.